git-svn-id: http://svn.code.sf.net/p/fbide/code/trunk@249 957c6b5c-1c3a-0410-895f-c76cfc11fbc7

+++ fbide-vs/Sdk/DocManager.cpp
@@ -109,7 +109,7 @@ struct TheDocManager : DocManager
         wxWindowUpdateLocker uiLock(GET_FRAME());
 
         // Open the files
-        for (int i = 0; i < paths.Count(); i++)
+        for (size_t i = 0; i < paths.Count(); i++)
         {
             auto doc = tm->CreateFromFile(paths[i]);
             if (doc)
